refactor browser support classes.
