+++ src/tools/browser/BrowserManager.ts
