+++ src/tools/browser/browser-manager.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, beforeEach, afterEach } from "vitest";
-import { BrowserManager } from "./browser-manager.js";
+import { BrowserManager } from "./BrowserManager.js";
 import { BrowserError, BrowserErrorCode } from "./types.js";
 
 describe("BrowserManager", () => {
