+++ src/tools/browser/element-state.test.ts
@@ -1,6 +1,6 @@
 /* eslint-disable max-lines-per-function */
 import { describe, it, expect, beforeAll, afterAll, beforeEach } from "vitest";
-import { BrowserManager } from "./browser-manager.js";
+import { BrowserManager } from "./BrowserManager.js";
 import { BrowserSession } from "./types.js";
 
 describe("Element State Tests", () => {
