+++ src/tools/browser/wait-behavior.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, beforeAll, afterAll, beforeEach } from "vitest";
-import { BrowserManager } from "./browser-manager.js";
+import { BrowserManager } from "./BrowserManager.js";
 import { BrowserSession } from "./types.js";
 
 describe("Wait Behavior Tests", () => {
